Reorganize web backend folder structure. Add frontend app skeleton
