+++ web/app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AppConfig(AppConfig):
+    name = 'app'
