+++ web/app/static/app/charts.js
