+++ web/app/static/app/dark.css
