+++ web/frontend/src/Oatmeal.jsx
@@ -1,0 +1,14 @@
+import './Oatmeal.css';
+
+function Oatmeal() {
+  return (
+    <div className="oatmeal">
+      <header className="oatmeal-header">
+        <h1>The Real Oatmeal</h1>
+        <p>Website rewrite in progress</p>
+      </header>
+    </div>
+  );
+}
+
+export default Oatmeal;
