+++ web/frontend/src/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Oatmeal from './Oatmeal';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Oatmeal />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
